// src/app/models/utilisateur.ts
import { Parent } from '../models/parent'; 
export class Utilisateur {

 constructor() {
    this.parent = new Parent();
  }

  id: number;
  username: string;
  password: string;
  email: string;
  grant_type: string;
  client_id: string;
  client_secret: string;
  profile_url: string;
  token: string;

  nom: string;
  prenom: string;
  dateNais: any;
  profession: string;
  telephone: string;
  adresse: string;
  matricule: string;
  role: string;
  lieuNais: string;
  sexe: string;
  handicape: false;
  enabled: boolean;
  parent?: any;
  roles=[];
}